requirements.txt
#Libraries Required:

matplotlib.pyplot
tensorflow
numpy
PIL
efficientnet.keras
pandas
os
scikit-learn
